% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_plot.R
\name{brand_plot}
\alias{brand_plot}
\title{Function to apply SP branding to the plot}
\usage{
brand_plot(plot_name)
}
\arguments{
\item{plot_name}{The variable name of the plot created that you intend to apply branding too.}
}
\value{
branded ggplot.
}
\description{
Function to apply SP branding to the plot
}
\examples{
sp <- ggplot2::ggplot(ggplot2::diamonds, ggplot2::aes(x=carat, y=price, color=cut)) +
      ggplot2::geom_point() +
      sp_theme() +
      ggplot2::labs(title="This is my title",
           subtitle = "This is my subtitle",
           caption = "Source: this is my data source",
           x="Carat",
           y="Price")
brand_plot(plot_name = sp)

}
\keyword{brand_plot}
