% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_datasets.R
\docType{data}
\name{DataSets-sp}
\alias{DataSets-sp}
\alias{Loadings2}
\alias{Loadings3}
\alias{Loadings4}
\alias{Scores2}
\alias{Scores3}
\alias{Scores4}
\alias{Spectra2}
\alias{Spectra3}
\alias{Spectra4}
\title{[+] Datasets of simulated spectroscopic data}
\format{\code{\link[=hyperSpec-class]{hyperSpec}} objects with spectroscopic
data and additional variables.

Additional variables for \code{Spectra..} and \code{Scores..}:

\describe{
  \item{class}{A factor variable with 4 classes.}
  \item{gr}{A factor variable with 3 classes.}
}

Additional variables for \code{Loadings..}:
\describe{
  \item{PeakAt}{Position of components top peak.}
  \item{cNames}{Names of components.}
  \item{order.of.rows}{Original order of components before sorting.}
}}
\source{
Artificially generated in R.
}
\usage{
Loadings2

Loadings3

Loadings4

Scores2

Scores3

Scores4

Spectra2

Spectra3

Spectra4
}
\description{
Datasets that simulate spectroscopic data (\code{Spectra..}
series - a mixture of spectral components)
and spectroscopic decomposition data (\code{Loadings..} series - spectra
of spectroscopic components and \code{Scores..} series - amplitudes
of spectroscopic componens).
}
\author{
Vilmantas Gegzna
}
\keyword{datasets}
