% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr2latex_.R
\name{expr2latex_}
\alias{expr2latex_}
\title{[!.] Translate `plotmath` expressions to LaTeX and remove leading and tailing quotes}
\usage{
expr2latex_(expr)
}
\arguments{
\item{expr}{an \R object of \code{\link{class}} \code{expression} or
    \code{language}, typically as from \code{\link{quote}(...)}.}
}
\value{
A \link[base]{character} string with the LaTeX expression
corresponding to "R graphics annotation" expression \code{expr}.
}
\description{
\code{\link[simsalapar]{expr2latex}} translates a "R graphics annotation"
expression to the corresponding LaTeX one.\cr
\code{expr2latex_} additionally removes leading and tailing  double quotes (").
}
\examples{

library(spPlot)
library(simsalapar)
library(hyperSpec)

expr2latex_(quote(N[sim]))


labels(chondro)$.wavelength
  ## expression(Delta * tilde(nu)/cm^-1)

labels(chondro)$.wavelength  \%>\% expr2latex
  ## [1] "Delta * tilde(nu)/cm^-1"

labels(chondro)$.wavelength  \%>\% expr2latex_
  ## [1] "Delta * tilde(nu)/cm^-1"



labels(chondro)$spc
  ## expression("I / a.u.")

labels(chondro)$spc  \%>\% expr2latex
## [1] "\\"I / a.u.\\""

labels(chondro)$spc  \%>\% expr2latex_
## [1] "I / a.u."
}
\seealso{
\code{\link[simsalapar]{expr2latex}} - translate 'plotmath'
expressions to LaTeX.\cr
\code{\link{label_expr2text}} remove expresions from labels of \code{ggplot2},
\code{hyperSpec} and \code{plotly} objects.

Other \pkg{spPlot} utilities: \code{\link{label_expr2text}}
}
\author{
Vilmantas Gegzna
}
