% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_wl.R
\name{facet_wl}
\alias{facet_wl}
\title{[!] Control facetting parameters in `ggplot` of `hyperSpec` object}
\usage{
facet_wl(rows = ".", space = "free", scales = "free", ...)
}
\arguments{
\item{rows}{Name of factor variable (quoted string) to be used for making
rows of facets. The dot "." (default) is used to indicate that there should
be no faceting on this dimension.}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.
If \code{"free_y"} their height will be proportional to the length of the
y scale; if \code{"free_x"} their width will be proportional to the
length of the x scale; or if \code{"free"} both height and width will
vary.  This setting has no effect unless the appropriate scales also vary.}

\item{scales}{Are scales shared across all facets (the default,
\code{"fixed"}), or do they vary across rows (\code{"free_x"}),
columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{...}{Further parameters to be passed to \code{\link[ggplot2]{facet_grid}}.}
}
\value{
Updated \pkg{ggplot2} object of class \code{\link[ggplot2]{ggplot}}.
}
\description{
Control facetting parameters in plot made with function \code{ggplot}
of \code{\link[=hyperSpec-class]{hyperSpec}} (\code{ggplot.hyperSpec} method)
in which parameter \code{wl.range} is explicitly specified.
}
\details{
This function is a wrapper for \code{\link[ggplot2]{facet_grid}} and
\code{facet_grid(.~.wl.range)} can be used instead of \code{facet_wl(".")}.
}
\examples{
library(spPlot)

p <- ggplot(Spectra2, wl.range = c(min ~ 350, 500~600)) +
     geom_line(aes(color = gr))

p
p + facet_wl(space = "fixed")
p + facet_wl("class")
p + facet_wl("gr")

# Alternative to `facet_wl()` is:
p + facet_grid(. ~ .wl.range, scales = "free")

\donttest{\dontrun{
# If `wl.range` is not specified, ERROR message is displayed:
p2 <- ggplot(Spectra2) + geom_line(aes(color = gr))
p2 + facet_wl()
}}


}
\seealso{
Function \code{\link[ggplot2]{facet_grid}} from package \pkg{ggplot2}.

Other \pkg{spPlot} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{ggLims}()},
\code{\link{gg_palette}()},
\code{\link{ggplot.hyperSpec}()},
\code{\link{label_expr2text}()}

Other \pkg{spPlot} functions for \pkg{ggplot2}: 
\code{\link{ggLims}()},
\code{\link{gg_palette}()},
\code{\link{ggplot.hyperSpec}()},
\code{\link{nTick}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spPlot} functions for \pkg{ggplot2}}
\concept{\pkg{spPlot} functions for spectroscopy and \pkg{hyperSpec}}
