% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_barplot.R
\name{gg_barplot}
\alias{gg_barplot}
\alias{gg_barplot.hyperSpec}
\alias{gg_barplot.default}
\title{A barplot with counts above bars.}
\usage{
gg_barplot(x, data, alpha = 1, color = "black", drop_NA = FALSE, ...)

\method{gg_barplot}{hyperSpec}(x, data, alpha = 1, color = "black", drop_NA = FALSE, ...)

\method{gg_barplot}{default}(x, data, alpha = 1, color = "black", drop_NA = FALSE, ...)
}
\arguments{
\item{x}{(string) A name of factor variable in the data frame.}

\item{data}{A data frame.}

\item{alpha}{(number from 0 to 1) Level of transperency.
Default is 1 (not transperent).}

\item{color}{(string) Edge color of bars.}

\item{drop_NA}{(logical) Should NA values be dropped?
Default is \code{FALSE}.}

\item{...}{(not used yet)}
}
\description{
A simple `ggplot2` barplot with counts above bars.
}
\examples{
#' library(spPlot)

#`data` is a data frame:
gg_barplot("cyl", mtcars)


# data` is a `hyperSpec` object:
gg_barplot("class", Spectra2)

}
