% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_palette.R, R/palette2vec.R
\name{gg_palette}
\alias{gg_palette}
\alias{palette2vec}
\title{[!.] Add color palette to `ggplot`}
\usage{
gg_palette(obj, palette = NULL, name = NULL, ...)

palette2vec(palette)
}
\arguments{
\item{obj}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{palette}{Color palette. Eithar a named or unnamed vector with colors,
or data frame with variables \code{$colors} and \code{labels} (data frame is
usually generated with function \code{\link{hyGet_palette0}}).}

\item{name}{the main title of the legend.}

\item{...}{further parameters to be passed to
\code{\link[ggplot2]{scale_color_manual}} and
\code{\link[ggplot2]{scale_fill_manual}}}
}
\value{
Updated \pkg{ggplot2} object of class \code{\link[ggplot2]{ggplot}}.
}
\description{
[!.] Add color palette to `ggplot`
}
\details{
\code{gg_palette} adds color palette to \code{ggplot} object. It wraps two
functions \code{\link[ggplot2]{scale_color_manual}} and
\code{\link[ggplot2]{scale_fill_manual}}.

\code{palette2vec} converts dataframe with variables \code{$colors}
and \code{$labels} to a named vector, were \code{$labels} are the names and
\code{$colors} are the values.The data frame is usually generated with function
\code{\link{hyGet_palette0}}.
}
\examples{
## NO EXAMPLES YET ##

}
\seealso{
Other \pkg{spPlot} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{facet_wl}},
  \code{\link{ggLims}}, \code{\link{ggplot.hyperSpec}},
  \code{\link{label_expr2text}}

Other \pkg{spPlot} functions for \pkg{ggplot2}: \code{\link{facet_wl}},
  \code{\link{ggLims}}, \code{\link{ggplot.hyperSpec}},
  \code{\link{nTick}}
}
\author{
Vilmantas Gegzna
}
