% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.hyperSpec.R
\name{ggplot.hyperSpec}
\alias{ggplot.hyperSpec}
\title{[!] \code{ggplot} method for class \code{hyperSpec}}
\usage{
\method{ggplot}{hyperSpec}(obj, mapping = aes(), ..., wl.range = NULL,
  palette = NULL, format = "\%g", environment)
}
\arguments{
\item{obj}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
Default values are \code{aes_string(x = ".wavelength",
y = "spc", group = ".rownames")}. Provided pameters update and modify this
list. Other not specified parameters must be suppled in each layer added
to the plot.}

\item{...}{further arguments to be passed to function
\code{\link[ggplot2]{ggplot}}.}

\item{wl.range}{wavelength ranges to plot. The ranges will be plotted
as separate columns of facets. Examples in \code{\link[hyperSpec]{wl2i}}
show how to select wavelength ranges of \code{hyperSpec} object \cr
 If \code{wl.range} is not \code{NULL}
function \code{\link{facet_wl}} may be used to control facetting
parameters. Otherwise standard \pkg{ggplot2}  functions, such as
\code{\link[ggplot2]{facet_grid}} or
\code{\link[ggplot2]{facet_wrap}} should be used if facetting is needed.}

\item{palette}{[Expreimental!!!] Color palete for \code{color} and \code{fill}.}

\item{format}{A string with format to be passed to \code{\link[base]{sprintf}}.
This string is used to format text in labels of wavelength intervals, when
\code{wl.range} is specified. Default format is \code{'\%g'}.}

\item{environment}{Not used yet.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
       More details in package \pkg{ggplot2}.
}
\description{
Method to create a new \code{ggplot} object for class \code{hyperSpec}.
More details in \code{\link[ggplot2]{ggplot}}.
}
\examples{
library(spPlot)
library(spHelper)
data(Spectra2)

ggplot(Spectra2) + geom_line()
ggplot(Spectra2, aes(color=class)) + geom_line()
ggplot(Spectra2) + geom_line(color = "red", alpha = 0.1) + theme_bw()


ggplot(Spectra2, wl.range = c(min ~ 320, 500~600, 700 ~ max)) +
     geom_line(aes(color = gr))

Sp2 <- hyAdd_color(Spectra2, "gr", c("green","green3","skyblue"))
ggplot(Sp2, palette = TRUE) + geom_line()

ggplot(Sp2, aes(color = gr), palette = c("blue","gold4","tan")) + geom_line()

ggplot(Sp2) + geom_line(aes(color = I(.color)))


library(ggspectra)
ggplot(Sp2[1]) + stat_color(geom = "bar")+ scale_color_identity()
}
\seealso{
\code{\link[hyperSpec]{qplotspc}} in package \pkg{hyperSpec}.\cr
\code{\link[ggplot2]{ggplot}} in package \pkg{ggplot2}.

Other \pkg{spPlot} functions for \pkg{ggplot2}: \code{\link{facet_wl}},
  \code{\link{ggLims}}, \code{\link{gg_palette}},
  \code{\link{nTick}}

Other \pkg{spPlot} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{facet_wl}},
  \code{\link{ggLims}}, \code{\link{gg_palette}},
  \code{\link{label_expr2text}}
}

