% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_ranges.R
\name{highlight_ranges}
\alias{highlight_ranges}
\title{Highlight wavelength ranges in a ggplot2 plot of hyperSpec object}
\usage{
highlight_ranges(
  zones,
  fill = RColorBrewer::brewer.pal(9, "Set1")[1:length(zones)],
  alpha = 0.2
)
}
\arguments{
\item{zones}{formula or list of formulas, that indicate which  zones of
wavelength axis should be highlighted e.g., 460~800.}

\item{fill}{vector of rectangle fill colors to be used for each zone;}

\item{alpha}{(number from 0 to 1) Level of transperency.
Default is 1 (not transperent).}
}
\description{
Highlight wavelength ranges in a ggplot2 plot of hyperSpec object
}
\details{
\code{highlight_ranges()} Can be added to ggplot object using "+" operator.
}
\examples{
gg <- ggplot(Spectra2) + geom_line()

gg + highlight_ranges(400~500)

zones <- c(400~460, 545~670)
gg + highlight_ranges(zones)
}
