% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Join.R
\name{Join}
\alias{Join}
\title{[!] Merge objects to `knitr_container`}
\usage{
Join(..., list = NULL)
}
\arguments{
\item{...}{objects to be merged.}

\item{list}{objects to be merged provided as a list.}
}
\value{
knitr_container
}
\description{
\code{Join} converts objects to \code{knitr_container} objects
(method "as is") and \bold{merges} them to one container.
}
\examples{

# Find examples in link `knitr_container-class`

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \code{knitr_container} functions: \code{\link{add_as_section}},
  \code{\link{add_as_}},
  \code{\link{knitr_container-class}},
  \code{\link{knitr_container}},
  \code{\link{print_objects}}, \code{\link{print}},
  \code{\link{summary}}
}

