% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_container-class.R
\docType{class}
\name{knitr_container-class}
\alias{knitr_container-class}
\title{S3 class "\code{knitr_container}"}
\source{
Some ideas and pieces of code for \code{knitr_container} functions are taken from
 \href{https://github.com/ropensci/plotly/issues/273#issuecomment-217982315}{this answer} on github.com.
}
\description{
\code{knitr_container} is an object for collecting other objects,
 such as section headings, text (as character vectors),
 \code{ggplot} and \code{plotly} objects, \code{pander} tables, etc.
 in an R Markdown file and print them in one function call. The advantage
 of such printing is obvious in those cases, where otherwise the object are
 not printed, e.g., from inside \code{for} loop when \code{HTML} output
 is desired.
}
\examples{

library(spPlot)

plotly_obj <- plot_ly(economics, x = date, y = uempmed, type = "scatter",
             showlegend = FALSE)

ggplot_obj <- qplot(mpg, wt, data = mtcars, colour = cyl)


container <- knitr_container()
container <- add_as_section(container, "Plots")

container <- add_as_section(container, "Add plotly", level = 2)
container <- add_as_plotly(container, plotly_obj)

container <- add_as_section(container, "Add ggplot as plotly", level = 2)
container <- add_as_plotly(container, ggplot_obj)

container <- add_as_section(container, "Add ggplot", level = 2)
container <- add_as_is(container, ggplot_obj)

container <- add_as_section(container, "Pander and text", level = 1)

container <- add_as_section(container, "As pander", level = 2)
container <- add_as_pander(container, summary(mtcars))

container <- add_as_section(container, "As text", level = 2)
container <- add_as_text(container, summary(mtcars))

container <- add_as_section(container, "As is", level = 2)
container <- add_as_is(container, summary(mtcars))

container <- add_as_is(container, plotly_obj)

summary(container)
print(container)
print_objects(container)

is.knitr_container(container)
is.knitr_container(ggplot_obj)

as.knitr_container(ggplot_obj)

class(container)


Join(container, container)
Join(container, ggplot_obj)
Join(ggplot_obj)
}
\author{
Vilmantas Gegzna
}
\seealso{
Examples in vignette `v3_knitr_container`.

Other \code{knitr_container} functions: \code{\link{Join}},
  \code{\link{add_as_}}, \code{\link{knitr_container}},
  \code{\link{print_objects}}, \code{\link{print}},
  \code{\link{summary}}
}
\keyword{classes}

