% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_container.R
\name{knitr_container}
\alias{as.knitr_container}
\alias{is.knitr_container}
\alias{knitr_container}
\title{Create, test and convert to `knitr_container` object}
\usage{
knitr_container()

is.knitr_container(obj)

as.knitr_container(obj = NULL, ...)
}
\arguments{
\item{obj}{object to be tested or converted to \code{knitr_container} object.}

\item{...}{arguments to be passed to function \code{knitr_container}.}
}
\description{
\code{knitr_container()} - creates a new empty \code{knitr_container} object.

\code{is.knitr_container()} - tests if class of an object is object
                            \code{knitr_container()}.

\code{as.knitr_container()} - converts an object to \code{knitr_container}.
 More specifically:\cr
for list - only the class attribute is added,\cr
for other objects - firstly an object is inctluded into list by
\code{list(object)} and then class attribute is added;\cr
for \code{knitr_container} objects - they are returned as-is.
}
\examples{

# Find examples in link `knitr_container-class`

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \code{knitr_container} functions: \code{\link{Join}},
  \code{\link{add_as_section}}, \code{\link{add_as_}},
  \code{\link{knitr_container-class}},
  \code{\link{print_objects}}, \code{\link{print}},
  \code{\link{summary}}
}

