% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_expr2text.R
\name{label_expr2text}
\alias{label_expr2text}
\alias{label_expr2text.ggplot}
\alias{label_expr2text.hyperSpec}
\alias{label_expr2text.plotly_hash}
\alias{label_expr2text.plotly}
\title{[!] Convert 'plotmath' expresions in labels to text}
\usage{
label_expr2text(obj, FUN = expr2latex_, ...)

\method{label_expr2text}{ggplot}(obj, FUN = expr2latex_, ...)

\method{label_expr2text}{hyperSpec}(obj, FUN = expr2latex_, ...)

\method{label_expr2text}{plotly_hash}(obj, FUN = expr2latex_, ...)

\method{label_expr2text}{plotly}(obj, FUN = expr2latex_, ...)
}
\arguments{
\item{obj}{Object of approptiate class.}

\item{FUN}{Function that translates expressions in labels to text or
other kind of expressions to be applied, e.g., simsalapar::expr2latex,
simsalapar::escapeLatex, expr2latex_.}

\item{...}{For generic use.}
}
\value{
Updated object with expresions in its labels converted to strings.
}
\description{
[!] Convert 'plotmath' expresions in labels to text
}
\examples{

library(spPlot)
library(hyperSpec)
library(plotly)

# ===== for hyperSpec object ===============================================

labels(chondro)
labels(label_expr2text(chondro))

\dontrun{\donttest{
# without `label_expr2text` labels with expressions are distorted or not plotted at all:

chondro[1:10,,] \%>\% qplotspc \%>\% ggplotly
}}
# If function `label_expr2text` is applied:
chondro[1:10,,] \%>\% label_expr2text  \%>\% qplotspc \%>\% ggplotly


# ===== for ggplot2 object =================================================

chondro[1:10,,] \%>\% qplotspc \%>\% label_expr2text \%>\% ggplotly


# ===== for plotly object ==================================================

chondro[1:10,,]  \%>\% qplotspc \%>\%  ggplotly \%>\% label_expr2text
\dontrun{\donttest{
## Does not work with newer version of `plotly`:
chondro[1:10,,]  \%>\% qplotspc \%>\%  ggplotly \%>\% plotly_build \%>\% label_expr2text
}}
# ==========================================================================
}
\seealso{
\code{\link{expr2latex_}}

Other \pkg{spPlot} utilities: \code{\link{expr2latex_}}

Other \pkg{spPlot} functions for \pkg{plotly}: \code{\link{plotly_annotation_rm}},
  \code{\link{plotly_modify_hover}},
  \code{\link{plotly_modify_legend}},
  \code{\link{plotly_tidy}}

Other \pkg{spPlot} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{facet_wl}},
  \code{\link{ggLims}}, \code{\link{gg_palette}},
  \code{\link{ggplot.hyperSpec}}
}
\author{
Vilmantas Gegzna
}
