% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nTick.R
\name{nTick}
\alias{nTick}
\alias{nTick_x}
\alias{nTick_y}
\title{[+] Control number of ticks in \code{ggplot2} plots with continuous scale}
\usage{
nTick_x(n = 2, min.n = 2, ..., pretty.args = list())

nTick_y(n = 2, min.n = 2, ..., pretty.args = list())
}
\arguments{
\item{n}{An Integer giving the desired number of intervals. Non-integer
values are rounded down.}

\item{min.n}{nonnegative integer giving the minimal number of intervals.
If \code{min.n == 0}, \code{pretty(.)} may return a single value.}

\item{...}{Further parameters to be passed to appropriate function
which is either \code{scale_x_continuous} for \code{nTick_x} or
\code{scale_y_continuous} for \code{nTick_y}.}

\item{pretty.args}{Further parameters to be passed to function \code{pretty}.}
}
\description{
Convenience function to control number of ticks in \code{ggplot2}
 plots with continuous scale.
}
\examples{

library(spPlot)
library(ggplot2)

# Make a plot but do not print
p <- qplot_spStat(Spectra2, "gr", mean, All.linetype = "solid") + facet_grid(.~gr)

#Print the plot
p

# Correct number of ticks:
p + nTick_x(2)
p + nTick_x(2) + nTick_y(8)

}
\author{
Vilmantas Gegzna
}
\seealso{
Pretty breakpoints: \code{\link[base]{pretty}}.

Other \pkg{spPlot} functions for \pkg{ggplot2}: \code{\link{facet_wl}},
  \code{\link{ggLims}}, \code{\link{gg_palette}},
  \code{\link{ggplot.hyperSpec}}
}

