% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_datasets.R
\docType{data}
\name{palette_PAP}
\alias{palette_PAP}
\title{[+] Color palette}
\format{An object of class \code{character} of length 8.}
\usage{
palette_PAP
}
\description{
Color palette, witch is \code{RColorBrewer::brewer.pal(8, "Set1")},
where colors go in a different order.
}
\details{
\code{palette_PAP <- c("#4DAF4A", "#377EB8", "#F781BF","#FF7F00",  "#E41A1C", "#A65628", "#984EA3",  "#FFFF33")}
}
\keyword{datasets}
