% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_annotation_rm.R
\name{plotly_annotation_rm}
\alias{plotly_annotation_list}
\alias{plotly_annotation_n}
\alias{plotly_annotation_rm}
\alias{plotly_annotation_text}
\title{[!] Manage annotations in a \code{plotly} object}
\usage{
plotly_annotation_rm(obj, which = NULL)

plotly_annotation_list(obj, which = NULL)

plotly_annotation_n(obj)

plotly_annotation_text(obj)
}
\arguments{
\item{obj}{a \pkg{plotly} object (created with function \code{plot_ly},
\code{ggplotly}, \code{subplot} or similar).}

\item{which}{Numeric vector with numbers of annotations to remove.}
}
\value{
The functions described here do the following actions to \code{plotly}
 object:\cr\cr
\code{plotly_annotation_rm}() removes annotations and returns
updated \code{plotly} object;\cr\cr
\code{plotly_annotation_n}() returns the number of annotations in the object;\cr\cr
\code{plotly_annotation_list}() returns a list of annotation parameters;\cr\cr
\code{plotly_annotation_text}() returns a list with text of each annotation.
}
\description{
Manage annotations of \code{plotly} object.
}
\examples{
library(spPlot)
library(plotly)

# Create plotly object 'obj':
iris$Species2 <- sample(iris$Species)
p <- ggplot(iris, aes(x = Sepal.Length,
                         y = Sepal.Width))
obj <- subplot(p + geom_line(aes(color = Species)),
               p + geom_point(aes(shape = Species2)))

# Print the plot:
obj

# Get parameters of annotations:
plotly_annotation_list(obj)
plotly_annotation_n(obj)
plotly_annotation_text(obj)


# The same plot with all annotations removed:
plotly_annotation_rm(obj)

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spPlot} functions for \pkg{plotly}: \code{\link{label_expr2text}},
  \code{\link{plotly_modify_hover}},
  \code{\link{plotly_modify_legend}},
  \code{\link{plotly_tidy}}
}

