% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_modify_hover.R
\name{plotly_modify_hover}
\alias{plotly_modify_hover}
\title{[!] Modify hover in a \code{plotly} object}
\usage{
plotly_modify_hover(
  obj,
  new_text = list(),
  hovermode = NULL,
  hoverinfo = NULL,
  hoverformat = NULL
)
}
\arguments{
\item{obj}{a \pkg{plotly} object (created with function \code{plot_ly},
\code{ggplotly}, \code{subplot} or similar).}

\item{new_text}{Either \code{NULL}, if the text should not be updated, or
 \bold{named list} (or object convertible to named list) with
 with replaced and replacement text in such form: \cr
\code{list("OLD_text" = "REPLACEMENT_text",
 "OLD_text2" = "REPLACEMENT_text2",)}.\cr
  If not \code{NULL}, and not explicitly modified the default
  values are:\cr
\code{new_text <- list(
".wavelength" = "Wavelength",
"spc" = "Y axis value",
".rownames" = "Row",
".aggregate" = "Group")}. \cr}

\item{hovermode}{(enumerated: \code{"closest"} | \code{FALSE})\cr
Source:\url{https://plot.ly/r/reference/#layout-scene-hovermode}.\cr
If \code{NULL} (default) value of this option will not be changed.}

\item{hoverinfo}{[NOT IMPLEMENTED YET. Option has no effect.] (flaglist string)
 Any combination of "x", "y", "z", "text", "name" joined with a "+" OR "all" or "none". \cr
Examples: "x", "y", "x+y", "x+y+z", "all".
Determines which trace information appear on hover.\cr
 Source: \url{https://plot.ly/r/reference/#box-hoverinfo}.\cr
 If \code{NULL} (default) value of this option will not be changed.}

\item{hoverformat}{(string)  [NOT IMPLEMENTED YET. Option has no effect.]
Sets the hover text formatting rule for data values on this axis, using
the python/d3 number formatting language. See
\url{https://github.com/mbostock/d3/wiki/Formatting#numbers} or
\url{https://docs.python.org/release/3.1.3/library/string.html#formatspec}
 for more info.\cr
 Source: \url{https://plot.ly/r/reference/#layout-yaxis-hoverformat}.\cr
 If \code{NULL} (default) value of this option will not be changed.}
}
\value{
An updated \pkg{plotly} object (class \code{plotly}).
}
\description{
Function corrects text in a hover of  a \code{plotly} object.
}
\examples{
library(spPlot)
library(spHelper)
library(hyperSpec)
library(plotly)

# Create `plotly` object:
obj <- qplot_spStat(chondro,"clusters",mean)  \%>\%
       label_expr2text() \%>\%
       ggplotly()


# Then compare hover in this plot:
plotly_modify_hover(obj)

# With the hover in the original plot:
print(obj)

}
\seealso{
Online reference
\url{https://plot.ly/r/reference/#scatterternary-hoverinfo}; \cr
Functions \link[plotly]{layout}(), \link[plotly]{plot_ly}(),

Other \pkg{spPlot} functions for \pkg{plotly}: 
\code{\link{label_expr2text}()},
\code{\link{plotly_annotation_rm}()},
\code{\link{plotly_modify_legend}()},
\code{\link{plotly_tidy}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spPlot} functions for \pkg{plotly}}
