% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_modify_legend.R
\name{plotly_modify_legend}
\alias{plotly_modify_legend}
\title{[!] Modify legend of a \code{plotly} object}
\usage{
plotly_modify_legend(obj, showlegend = NULL, traceorder = NULL,
  bgcolor = NULL, bordercolor = NULL, borderwidth = NULL,
  fontcolor = NULL, fontfamily = NULL, fontsize = NULL, x = NULL,
  y = NULL, yanchor = NULL, xanchor = NULL, tracegroupgap = NULL,
  orientation = NULL, rm_repText = TRUE, unique.legend = TRUE, ...)
}
\arguments{
\item{obj}{a \pkg{plotly} object (created with function \code{plot_ly},
\code{ggplotly}, \code{subplot} or similar).}

\item{showlegend}{logical, which determins if will be shown.
If \code{NULL} (default) value of this option will not be changed.}

\item{traceorder}{Any combination of "reversed", "grouped" joined with a "+" OR "normal".
examples: "reversed", "grouped", "reversed+grouped", "normal"
Determines the order at which the legend items are displayed. If "normal",
 the items are displayed top-to-bottom in the same order as the input data.
 If "reversed", the items are displayed in the opposite order as "normal".
 If "grouped", the items are displayed in groups (when a trace `legendgroup`
 is provided). if "grouped+reversed", the items are displayed in the
 opposite order as "grouped". Described in
 \href{https://plot.ly/r/reference/#layout-legend}{R plotly reference}.}

\item{bgcolor}{Legend's background color. If \code{NULL} (default) value of this option will not be changed.}

\item{bordercolor}{Legend's border color. If \code{NULL} (default) value of this option will not be changed.}

\item{borderwidth}{Legend's border width. If \code{NULL} (default) value of this option will not be changed.}

\item{fontcolor}{Legend's font color.     If \code{NULL} (default) value of this option will not be changed.}

\item{fontfamily}{Legend's font family.   If \code{NULL} (default) value of this option will not be changed.}

\item{fontsize}{Legend's font size.     If \code{NULL} (default) value of this option will not be changed.}

\item{x}{Legend's x position.             If \code{NULL} (default) value of this option will not be changed.}

\item{y}{Legend's y position.             If \code{NULL} (default) value of this option will not be changed.}

\item{yanchor}{Described in \href{https://plot.ly/r/reference/#layout-legend}{R plotly reference}.}

\item{xanchor}{Described in \href{https://plot.ly/r/reference/#layout-legend}{R plotly reference}.}

\item{tracegroupgap}{Described in \href{https://plot.ly/r/reference/#layout-legend}{R plotly reference}.}

\item{orientation}{"v","h". If \code{NULL} (default) value of this option will not be changed.
Described in \href{https://plot.ly/r/reference/#layout-legend}{R plotly reference}.}

\item{rm_repText}{Logical. If \code{TRUE} (default), repeated text in
legend's label is removed.}

\item{unique.legend}{logical. If \code{TRUE} (default), only one legend per
legend group is shown. Result is visible if function
\code{\link[plotly]{subplot}} was applied.}

\item{...}{Further parameters to methods.}
}
\value{
An updated \pkg{plotly} object (class \code{plotly}).
}
\description{
Function allows modifying various legend options of object (plot) created
with package \pkg{plotly}.
}
\details{
If option \code{rm_repText} is \code{TRUE} (default), function corrects text
in legend of object  created with function \code{\link[plotly]{ggplotly}} so
that only relevant information is displayed.
For example, if the text in a label is \code{"(matrix,matrix,matrix)"}, it is
converted to shorter string \code{"matrix"}.
}
\examples{
library(spPlot)
library(spHelper)
library(hyperSpec)
library(plotly)

# Create `plotly` object:
obj <- qplot_spStat(chondro,"clusters",mean)  \%>\% ggplotly

# Then compare this plot:
plotly_modify_legend(obj)

# With the original one:
print(obj)

# More options
plotly_modify_legend(obj, traceorder = "reversed")
plotly_modify_legend(obj, showlegend = FALSE)



}
\author{
Vilmantas Gegzna
}
\seealso{
Online reference \url{https://plot.ly/r/reference/#layout-legend};\cr
Functions \link[plotly]{layout}(), \link[plotly]{plot_ly}(),

Other \pkg{spPlot} functions for \pkg{plotly}: \code{\link{label_expr2text}},
  \code{\link{plotly_annotation_rm}},
  \code{\link{plotly_modify_hover}},
  \code{\link{plotly_tidy}}
}

