% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_tidy.R
\name{plotly_tidy}
\alias{plotly_tidy}
\alias{plotly_tidy0}
\title{Tidy up `plotly` plot}
\usage{
plotly_tidy(obj, ...)

plotly_tidy0(obj)
}
\arguments{
\item{obj}{a \pkg{plotly} object (created with function \code{plot_ly},
\code{ggplotly}, \code{subplot} or similar).}

\item{...}{arguments to be passed to \code{\link{plotly_annotation_rm}}.}
}
\value{
An updated \pkg{plotly} object (class \code{plotly}).
}
\description{
Tidy up \code{plotly} object, created from \code{ggplot} object.
}
\details{
This function is a wrapper of following functions with default
 parameters applied in sequence as follows:
\code{\link{label_expr2text}},
\code{\link{plotly_modify_legend}},
\code{\link{plotly_modify_hover}},
\code{\link{plotly_annotation_rm}}. \cr
\code{plotly_tidy0} does first 3 steps but does not remove annotations.
}
\author{
Vilmantas Gegzna
}
\seealso{
\link[plotly]{layout}(),
          \link[plotly]{plot_ly}(),
          \link[plotly]{subplot}().

Other \pkg{spPlot} functions for \pkg{plotly}: \code{\link{label_expr2text}},
  \code{\link{plotly_annotation_rm}},
  \code{\link{plotly_modify_hover}},
  \code{\link{plotly_modify_legend}}
}

