% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_objects.R
\name{print_objects}
\alias{print_objects}
\title{[!] Print objects from \code{knitr_container} in R Markdown / \code{knitr} file}
\usage{
print_objects(container)
}
\arguments{
\item{container}{\code{knitr_container} object.}
}
\description{
Function takes every object from the \code{container}, prints it
 and attaches \code{html} dependencies to show objects
 (such as plotly htmlwidgets) correctly.
}
\examples{

# Find examples in link `knitr_container-class`

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \code{knitr_container} functions: \code{\link{Join}},
  \code{\link{add_as_}},
  \code{\link{knitr_container-class}},
  \code{\link{knitr_container}}, \code{\link{print}},
  \code{\link{summary}}
}

