% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_theme.R
\name{sp_theme}
\alias{sp_theme}
\title{This function will apply SP theme to any ggplot.}
\usage{
sp_theme()
}
\description{
This function will apply SP theme to any ggplot.
}
\examples{
ggplot2::ggplot(ggplot2::diamonds, ggplot2::aes(x=carat, y=price, color=cut)) +
      ggplot2::geom_point() +
      sp_theme() +
      ggplot2::labs(title="This is my title",
           subtitle = "This is my subtitle",
           caption = "Source: this is my data source",
           x="Carat",
           y="Price")
# + theme(<ADDITIONAL ARGUMENTS TO OVERIDE DEFAULT SET BY SP_THEME>)
# NOTE : If sp_theme is applied at end it will override earlier arguments passed in theme ().
}
\keyword{sp_theme}
