% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_spPlot_description.R
\docType{package}
\name{spPlot}
\alias{spPlot}
\alias{spPlot-package}
\title{Various ploting functions for `hyperSpec`, `ggplot2`, `plotly`, etc}
\description{
[+] - function is well described. \cr

[!] - a description is incomplete and needs revision.\cr\cr
}
\author{
Vilmantas Gegzna
}

