% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_1.R
\name{subplot_1}
\alias{subplot_1}
\alias{subplot_2}
\title{Wrapper functions for plotly::subplot}
\usage{
subplot_1(
  ...,
  title = "Spectra and their derivatives",
  titleX = T,
  titleY = T,
  shareY = T,
  nrows = 1
)

subplot_2(
  ...,
  title = "Spectra and their derivatives",
  titleX = T,
  titleY = T,
  shareX = T,
  shareY = T,
  nrows = 2
)
}
\arguments{
\item{...}{inputs to be passed to \code{\link[plotly]{subplot}}.}

\item{title}{(string) the main title of the plot.}

\item{titleX}{should x-axis titles be retained?}

\item{titleY}{should y-axis titles be retained?}

\item{shareY}{should the y-axis be shared amongst the subplots?}

\item{nrows}{number of rows for laying out plots in a grid-like structure.
Only used if no domain is already specified.}

\item{shareX}{should the x-axis be shared amongst the subplots?}
}
\value{
\code{plotly} object
}
\description{
\code{subplot_1} puts figures in 1 row\cr
\code{subplot_2} puts figures in 2 rows\cr
}
\details{
subplot
}
\examples{

# ----- no examples yet -----

}
