% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.knitr_container.R
\name{summary}
\alias{summary}
\alias{summary.knitr_container}
\title{Print and summary methods for \code{knitr_container}}
\usage{
\method{summary}{knitr_container}(container, len = 30, units = "Kb")
}
\arguments{
\item{container}{\code{knitr_container} object.}

\item{len}{length of text (number of characters) to be shown in summary.
Default is 30.}

\item{units}{the units to be used in printing the size.  Allowed
    values are
    \code{"b"}, \code{"Kb"}, \code{"Mb"}, \code{"Gb"}, \code{"Tb"}, \code{"Pb"},
    \code{"B"}, \code{"KB"}, \code{"MB"}, \code{"GB"}, \code{"TB"}, \code{"PB"},
    \code{"KiB"}, \code{"MiB"}, \code{"GiB"}, \code{"TiB"},
    \code{"PiB"}, \code{"EiB"}, \code{"ZiB"}, \code{"YiB"},
    and \code{"auto"} (see \sQuote{Details}).  Can be abbreviated.}
}
\description{
Print summary of \code{knitr_container} object.
}
\examples{

# Find examples in link `knitr_container-class`

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \code{knitr_container} functions: \code{\link{Join}},
  \code{\link{add_as_}},
  \code{\link{knitr_container-class}},
  \code{\link{knitr_container}},
  \code{\link{print_objects}}, \code{\link{print}}
}

