% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_as_.R
\name{add_as_}
\alias{add_as_}
\alias{add_as_is}
\alias{add_as_pander}
\alias{add_as_plotly_widget}
\alias{add_as_text}
\title{Add object to \code{knitr_container}}
\usage{
add_as_is(container, obj)

add_as_text(container = NULL, obj)

add_as_plotly_widget(container = NULL, obj)

add_as_pander(container = NULL, obj, ...)
}
\arguments{
\item{container}{\code{knitr_container} object.}

\item{obj}{An object to be added to the \code{container}.}

\item{...}{Options to be passed to \code{\link[pander]{pander}}.}
}
\description{
Functions to transform and add objects to \code{knitr_container}.
Objects, that can be included in R \code{\link[base]{list}}, can also be
included in \code{knitr_container}.

\code{\link{add_as_section}()} subfamily functions have separate description page.
}
\details{
Following functions converts and formats object \code{obj} sch that it could
be appropriatly printed with function \code{\link{print_objects}}().

\code{add_as_is()} includes object \code{obj} in the \code{container} without
transformation (\bold{"as is"}). Function \code{print_objects} will print it
using regular \code{print} function. Note that in R Markdaown \code{Rmd}
file \code{knitr} chunk option \code{results='asis'} may distort the
"beautiful" formatting of the printed object. This function is appropriate
to ingludde \pkg{ggplot2} plots, if they have to be displayed as \code{gg}
plots and not \code{plotly} plots.

\code{add_as_text()} converts \code{obj} to \code{\link[base]{character}},
 formats as \bold{text} and includes it in the \code{container}. Function
 \code{print_objects} will print it as text.

\code{\link{add_as_section}()} converts \code{obj} to \code{\link[base]{character}},
 formats it as a \bold{heading of section} and includes it in the \code{container}.
 Function \code{print_objects} will print it as text.

\code{add_as_plotly_widget()} converts \pkg{plotly} and \pkg{ggplot2} objects to
plotly htmlwidget (details in \code{\link[plotly]{as.widget}}) and includes
it in the \code{container}. Function \code{print_objects} will print it as
plotly htmlwidget and attach
\code{html} dependencies.

\code{add_as_pander()} formats supported types of \code{obj} with An R Pandoc
 Writer's function \code{\link[pander]{pander}} and includes it in the
 \code{container}. Function \code{print_objects} will print the object as text.
}
\examples{
# Find examples in link `knitr_container-class`

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \code{knitr_container} functions: \code{\link{Join}},
  \code{\link{add_as_section}},
  \code{\link{knitr_container-class}},
  \code{\link{knitr_container}},
  \code{\link{print_objects}}, \code{\link{print}},
  \code{\link{summary}}
}

