% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLims.R
\name{ggLims}
\alias{get_ggLims}
\alias{ggLims}
\alias{ggLims.gg}
\alias{ggLims.numeric}
\alias{set_ggLims}
\title{[!] Convenience functions to get and set the axis limits in ggplot2 object}
\usage{
ggLims(obj, which = "y", ...)

\method{ggLims}{gg}(obj, which = "y", ...)

\method{ggLims}{numeric}(value, which = "y", ...)

get_ggLims(obj, which = "y", ...)

set_ggLims(value, which = "y", ...)
}
\arguments{
\item{obj}{A ggplot object.}

\item{which}{(A string) either "x", "y" (default) or "xy" which
indicates axis if inerest.}

\item{...}{For generic use.}

\item{value}{Values of limits: either vector of \bold{2 values} (min and max)
for axis indicated in \code{which} or vector of \bold{4 values}
(x min, x max, y min, y max) to be passed to function
\code{\link[ggplot2]{coord_cartesian}}. \emph{NOTE}, that this function behaves differently than \code{\link[ggplot2]{lims}}.}
}
\value{
One of the following depending on inputs: \cr
a) vector of ranges of \emph{one} axis as
\code{(min, max)},
b) list of limits for both axes as
 \code{list(x = (x min, x max), y = (y min, y max)}\cr
c) updated ggplot object
}
\description{
[!] Convenience functions to get and set the axis limits in ggplot2 object
}
\details{
\code{get_ggLims(obj, "y")} is a wrapper for \cr
\code{ggplot_build(obj)$panel$ranges[[1]]$y.range} \cr\cr

\code{set_ggLims(value, "y")} is a wrapper for \cr
\code{coord_cartesian(ylim = value)}
}
\examples{

obj <- qplot(mpg, wt, data = mtcars, geom = c("line","point"))

# Get y axis limits ==========================
get_ggLims(obj)
ggLims(obj)
ggLims(obj, "y")

ggLims(obj, "x")
ggLims(obj, "xy")

\donttest{
\dontrun{

# Set y  .axis limits ==========================

# Original plot:
obj

# Settim limmits this way:
obj + ggLims(c(0, 5))

# is the same as doing it this way:
obj + ggLims(c(0, 5),"y")

# and this way:
obj + coord_cartesian(ylim = c(0, 5))

# BUT differs from this way, where some lines are trimmed:
obj + ylim(c(0, 5))

}}

}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link[ggplot2]{lims}}

Other \pkg{spPlot} functions for \pkg{ggplot2}: \code{\link{facet_wl}},
  \code{\link{gg_palette}}, \code{\link{ggplot.hyperSpec}},
  \code{\link{nTick}}

Other \pkg{spPlot} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{facet_wl}},
  \code{\link{gg_palette}}, \code{\link{ggplot.hyperSpec}},
  \code{\link{label_expr2text}}
}

