% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_tidy.R
\name{plotly_tidy}
\alias{ggplotly_tidy}
\alias{ggplotly_tidy0}
\alias{plotly_tidy}
\alias{plotly_tidy0}
\title{Tidy up `plotly` plot}
\usage{
plotly_tidy(obj, title = NULL)

plotly_tidy0(obj, title = NULL, ...)

ggplotly_tidy(obj_gg = ggplot2::last_plot(), title = NULL, ...)

ggplotly_tidy0(obj_gg = ggplot2::last_plot(), title = NULL, ...)
}
\arguments{
\item{obj}{a \pkg{plotly} object (created with function \code{plot_ly},
\code{ggplotly}, \code{subplot} or similar).}

\item{title}{(string | NULL) the title of the plot.}

\item{...}{in \code{plotly_tidy} arguments to be passed to
     \code{\link{plotly_annotation_rm}}.\cr
in \code{ggplotly_tidy} arguments to be passed to
     \code{\link{plotly_annotation_rm}}.\cr}

\item{obj_gg}{\pkg{ggplot2} object.}
}
\value{
An updated \pkg{plotly} object (class \code{plotly}).
}
\description{
Functions \code{plotly_tidy} and \code{plotly_tidy0} tidy* up \code{plotly}
object, usually created using function \code{ggplotly}.\cr
Function \code{ggplotly_tidy0} converts \code{ggplot} object to a tity*
\code{plotly} plot.\cr
These are the convenience functions.\cr
* - see section "Details"
}
\details{
Function \code{plotly_tidy} is a wrapper of following functions
with default parameters applied in sequence as follows:\cr
\code{\link{label_expr2text}},\cr
\code{\link{plotly_modify_legend}},\cr
\code{\link{plotly_modify_hover}}\cr\cr

\code{plotly_tidy0} additionaly applies \code{\link{plotly_annotation_rm}(...)}
after the sequence described above.\cr\cr

\code{ggplotly_tidy} additionaly applies \code{\link{ggplotly}(...)}
before the sequence described above but does not remove annotations.
}
\author{
Vilmantas Gegzna
}
\seealso{
\link[plotly]{ggplotly}(),
          \link[plotly]{layout}(),
          \link[plotly]{plot_ly}(),
          \link[plotly]{subplot}(),
          \link[ggplot2]{ggplot}().

Other \pkg{spPlot} functions for \pkg{plotly}: \code{\link{label_expr2text}},
  \code{\link{plotly_annotation_rm}},
  \code{\link{plotly_modify_hover}},
  \code{\link{plotly_modify_legend}}
}

